requirements.txt
streamlit
langchain-core
langchain-community
langchain-huggingface
langchain-chroma
langchain-mistralai
langchain-tavily
langchain-text-splitters
chromadb
sentence-transformers
pymupdf
requests
python-dotenv
